% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_leaky_relu}
\alias{activation_leaky_relu}
\title{Leaky Rectified Linear Unit Activation Function}
\usage{
activation_leaky_relu(alpha = 0.01)
}
\arguments{
\item{alpha}{A non-negative numeric value controlling the slope for
negative inputs.}
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Creates a leaky rectified linear unit activation function object.
}
\examples{
act <- activation_leaky_relu(alpha = 0.01)
act$fn(c(-1, 0, 1))
}
