% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_linear}
\alias{activation_linear}
\title{Linear Activation Function}
\usage{
activation_linear()
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Creates a linear activation function object.
}
\examples{
act <- activation_linear()
act$fn(c(-1, 0, 1))
}
