% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_relu}
\alias{activation_relu}
\title{Rectified Linear Unit Activation Function}
\usage{
activation_relu()
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Creates a rectified linear unit activation function object.
}
\examples{
act <- activation_relu()
act$fn(c(-1, 0, 1))
}
\references{
Nair, V., and Hinton, G. E. (2010). Rectified Linear Units Improve
Restricted Boltzmann Machines. Proceedings of the 27th International
Conference on Machine Learning, 807--814.
}
