% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_sigmoid}
\alias{activation_sigmoid}
\title{Sigmoid Activation Function}
\usage{
activation_sigmoid()
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Creates a sigmoid activation function object.
}
\examples{
act <- activation_sigmoid()
act$fn(c(-1, 0, 1))
}
