% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_softmax}
\alias{activation_softmax}
\title{Softmax Activation Function}
\usage{
activation_softmax()
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Creates a softmax activation function object.
}
\examples{
act <- activation_softmax()
act$fn(c(1, 2, 3))
}
\references{
Bridle, J. S. (1990). Probabilistic Interpretation of Feedforward
Classification Network Outputs, with Relationships to Statistical Pattern
Recognition. In Neurocomputing: Algorithms, Architectures and Applications,
227--236. Springer.
}
