% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_tanh}
\alias{activation_tanh}
\title{Hyperbolic Tangent Activation Function}
\usage{
activation_tanh()
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Creates a hyperbolic tangent activation function object.
}
\examples{
act <- activation_tanh()
act$fn(c(-1, 0, 1))
}
