% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{as_activation}
\alias{as_activation}
\title{Convert Character Input to an Activation Object}
\usage{
as_activation(activation)
}
\arguments{
\item{activation}{A character string or an object of class
\code{"met_activation"}.}
}
\value{
An object of class \code{"met_activation"}.
}
\description{
Converts a character string such as \code{"relu"} into the corresponding
activation function object.
}
\examples{
as_activation("relu")
as_activation(activation_leaky_relu(alpha = 0.05))
}
