% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{as_loss}
\alias{as_loss}
\title{Convert Character Input to a Loss Object}
\usage{
as_loss(loss)
}
\arguments{
\item{loss}{A character string or an object of class \code{"met_loss"}.}
}
\value{
An object of class \code{"met_loss"}.
}
\description{
Converts a character string such as \code{"mse"} into the corresponding loss
function object.
}
\examples{
as_loss("mse")
as_loss(loss_huber(delta = 1.5))
}
