% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{as_metric}
\alias{as_metric}
\title{Convert Character Input to a Metric Object}
\usage{
as_metric(metric)
}
\arguments{
\item{metric}{A character string or an object of class \code{"met_metric"}.}
}
\value{
An object of class \code{"met_metric"}.
}
\description{
Converts a character string such as \code{"rmse"} into the corresponding metric
function object.
}
\examples{
as_metric("rmse")
as_metric(metric_accuracy())
}
