% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{as_metrics}
\alias{as_metrics}
\title{Convert Multiple Inputs to Metric Objects}
\usage{
as_metrics(metrics)
}
\arguments{
\item{metrics}{A character vector, a single metric object, or a list of
metric objects.}
}
\value{
A list of objects of class \code{"met_metric"}.
}
\description{
Converts a character vector or a list of metric objects into a list of
metric objects.
}
\examples{
as_metrics(c("rmse", "mae", "r2"))
as_metrics(list(metric_accuracy(), metric_f1()))
}
