% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{as_optimizer}
\alias{as_optimizer}
\title{Convert Character Input to an Optimizer Object}
\usage{
as_optimizer(optimizer)
}
\arguments{
\item{optimizer}{A character string or an object of class \code{"met_optimizer"}.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Converts a character string such as \code{"pso"} into the corresponding optimizer
object.
}
\examples{
as_optimizer("pso")
as_optimizer(optimizer_adam())
}
