% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_activations}
\alias{available_activations}
\title{List Available Activation Functions}
\usage{
available_activations()
}
\value{
A character vector of activation function names.
}
\description{
Returns the names of activation functions currently available in the
metANN package.
}
\examples{
available_activations()
}
