% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_gradient_optimizers}
\alias{available_gradient_optimizers}
\title{List Available Gradient-Based Optimizers}
\usage{
available_gradient_optimizers()
}
\value{
A character vector of gradient-based optimizer names.
}
\description{
Returns the names of gradient-based optimizer objects currently available
in the metANN package.
}
\examples{
available_gradient_optimizers()
}
