% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_losses}
\alias{available_losses}
\title{List Available Loss Functions}
\usage{
available_losses()
}
\value{
A character vector of loss function names.
}
\description{
Returns the names of loss functions currently available in the metANN
package.
}
\examples{
available_losses()
}
