% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_metaheuristics}
\alias{available_metaheuristics}
\title{List Available Metaheuristic Optimizers}
\usage{
available_metaheuristics()
}
\value{
A character vector of metaheuristic optimizer names.
}
\description{
Returns the names of metaheuristic optimization algorithms currently
available in the metANN package.
}
\examples{
available_metaheuristics()
}
