% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_metrics}
\alias{available_metrics}
\title{List Available Performance Metrics}
\usage{
available_metrics()
}
\value{
A character vector of metric names.
}
\description{
Returns the names of performance metrics currently available in the
metANN package.
}
\examples{
available_metrics()
}
