% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_optimize.R
\name{coef.met_optimize_result}
\alias{coef.met_optimize_result}
\title{Extract the Best Parameters from a metANN Optimization Result}
\usage{
\method{coef}{met_optimize_result}(object, ...)
}
\arguments{
\item{object}{A metANN optimization result object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A numeric vector containing the best solution found.
}
\description{
Extract the Best Parameters from a metANN Optimization Result
}
