% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metann.R
\name{coef.metann}
\alias{coef.metann}
\title{Extract Weights from a metANN Model}
\usage{
\method{coef}{metann}(object, ...)
}
\arguments{
\item{object}{A fitted metANN model.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A numeric vector of fitted network weights.
}
\description{
Extract Weights from a metANN Model
}
