% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_engine.R
\name{count_parameters}
\alias{count_parameters}
\title{Count the Number of Trainable Parameters in an MLP Architecture}
\usage{
count_parameters(architecture, input_dim = NULL)
}
\arguments{
\item{architecture}{An object created by \code{mlp_architecture()}.}

\item{input_dim}{Optional positive integer specifying the number of input
features. If \code{NULL}, \code{architecture$input_dim} is used.}
}
\value{
A positive integer giving the total number of parameters.
}
\description{
Computes the total number of weights and bias terms required by a
multilayer perceptron architecture.
}
\examples{
arch <- mlp_architecture(
  input_dim = 4,
  layers = list(
    dense_layer(5, activation = "relu"),
    dense_layer(1, activation = "linear")
  )
)
count_parameters(arch)
}
