% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{encode_classification_response}
\alias{encode_classification_response}
\title{Encode Classification Response}
\usage{
encode_classification_response(y)
}
\arguments{
\item{y}{Response vector.}
}
\value{
A list containing encoded response, class levels, and classification
type.
}
\description{
Internal helper for encoding binary and multi-class responses.
}
\keyword{internal}
