% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_optimize.R
\name{ga_optimize}
\alias{ga_optimize}
\title{Genetic Algorithm Engine}
\usage{
ga_optimize(fn, lower, upper, optimizer, verbose = TRUE, ...)
}
\arguments{
\item{fn}{Objective function.}

\item{lower}{Numeric vector of lower bounds.}

\item{upper}{Numeric vector of upper bounds.}

\item{optimizer}{A GA optimizer object.}

\item{verbose}{Logical. If \code{TRUE}, progress information is printed.}

\item{...}{Additional arguments passed to \code{fn}.}
}
\value{
An object of class \code{"met_optimize_result"}.
}
\description{
Internal implementation of a real-coded Genetic Algorithm.
}
\keyword{internal}
