% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{is_activation}
\alias{is_activation}
\title{Check Whether an Object is a metANN Activation}
\usage{
is_activation(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a metANN activation object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is a metANN Activation
}
\examples{
is_activation(activation_relu())
}
