% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{is_architecture}
\alias{is_architecture}
\title{Check Whether an Object is a metANN Architecture}
\usage{
is_architecture(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a metANN architecture object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is a metANN Architecture
}
\examples{
arch <- mlp_architecture(list(dense_layer(1)))
is_architecture(arch)
}
