% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{is_layer}
\alias{is_layer}
\title{Check Whether an Object is a metANN Layer}
\usage{
is_layer(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a metANN layer object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is a metANN Layer
}
\examples{
is_layer(dense_layer(5))
}
