% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{is_loss}
\alias{is_loss}
\title{Check Whether an Object is a metANN Loss}
\usage{
is_loss(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a metANN loss object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is a metANN Loss
}
\examples{
is_loss(loss_mse())
}
