% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{is_metric}
\alias{is_metric}
\title{Check Whether an Object is a metANN Metric}
\usage{
is_metric(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a metANN metric object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is a metANN Metric
}
\examples{
is_metric(metric_rmse())
}
