% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{is_mlp_architecture}
\alias{is_mlp_architecture}
\title{Check Whether an Object is an MLP Architecture}
\usage{
is_mlp_architecture(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is an MLP architecture object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is an MLP Architecture
}
\examples{
arch <- mlp_architecture(list(dense_layer(1)))
is_mlp_architecture(arch)
}
