% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{is_optimizer}
\alias{is_optimizer}
\title{Check Whether an Object is a metANN Optimizer}
\usage{
is_optimizer(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a metANN optimizer object; otherwise \code{FALSE}.
}
\description{
Check Whether an Object is a metANN Optimizer
}
\examples{
is_optimizer(optimizer_pso())
}
