% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_binary_crossentropy}
\alias{loss_binary_crossentropy}
\title{Binary Cross-Entropy Loss}
\usage{
loss_binary_crossentropy(epsilon = 1e-15)
}
\arguments{
\item{epsilon}{A small positive numeric value used for numerical stability.}
}
\value{
An object of class \code{"met_loss"}.
}
\description{
Creates a binary cross-entropy loss function object.
}
\examples{
loss <- loss_binary_crossentropy()
loss$fn(c(0, 1, 1), c(0.1, 0.8, 0.9))
}
\references{
Bridle, J. S. (1990). Probabilistic Interpretation of Feedforward
Classification Network Outputs, with Relationships to Statistical Pattern
Recognition. In Neurocomputing: Algorithms, Architectures and Applications,
227--236. Springer.
}
