% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_log_cosh}
\alias{loss_log_cosh}
\title{Log-Cosh Loss}
\usage{
loss_log_cosh()
}
\value{
An object of class \code{"met_loss"}.
}
\description{
Creates a log-cosh loss function object.
}
\examples{
loss <- loss_log_cosh()
loss$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
