% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_mae}
\alias{loss_mae}
\title{Mean Absolute Error Loss}
\usage{
loss_mae()
}
\value{
An object of class \code{"met_loss"}.
}
\description{
Creates a mean absolute error loss function object.
}
\examples{
loss <- loss_mae()
loss$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
