% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_mse}
\alias{loss_mse}
\title{Mean Squared Error Loss}
\usage{
loss_mse()
}
\value{
An object of class \code{"met_loss"}.
}
\description{
Creates a mean squared error loss function object.
}
\examples{
loss <- loss_mse()
loss$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
