% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_accuracy}
\alias{metric_accuracy}
\title{Accuracy Metric}
\usage{
metric_accuracy()
}
\value{
An object of class \code{"met_metric"}.
}
\description{
Creates an accuracy metric object for classification tasks.
}
\examples{
metric <- metric_accuracy()
metric$fn(c(0, 1, 1), c(0, 1, 0))
}
