% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_mae}
\alias{metric_mae}
\title{Mean Absolute Error Metric}
\usage{
metric_mae()
}
\value{
An object of class \code{"met_metric"}.
}
\description{
Creates a mean absolute error metric object.
}
\examples{
metric <- metric_mae()
metric$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
