% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_mse}
\alias{metric_mse}
\title{Mean Squared Error Metric}
\usage{
metric_mse()
}
\value{
An object of class \code{"met_metric"}.
}
\description{
Creates a mean squared error metric object.
}
\examples{
metric <- metric_mse()
metric$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
