% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_r2}
\alias{metric_r2}
\title{Coefficient of Determination Metric}
\usage{
metric_r2()
}
\value{
An object of class \code{"met_metric"}.
}
\description{
Creates an R-squared metric object.
}
\examples{
metric <- metric_r2()
metric$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
