% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_rmse}
\alias{metric_rmse}
\title{Root Mean Squared Error Metric}
\usage{
metric_rmse()
}
\value{
An object of class \code{"met_metric"}.
}
\description{
Creates a root mean squared error metric object.
}
\examples{
metric <- metric_rmse()
metric$fn(c(1, 2, 3), c(1.1, 1.9, 3.2))
}
