% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{mlp_architecture}
\alias{mlp_architecture}
\title{Create an MLP Architecture}
\usage{
mlp_architecture(layers, input_dim = NULL, name = "mlp")
}
\arguments{
\item{layers}{A list of dense layer objects created by \code{dense_layer()}.}

\item{input_dim}{Optional positive integer specifying the number of input
features. This can be left as \code{NULL} and inferred later from data.}

\item{name}{Optional character string specifying the architecture name.}
}
\value{
An object of class \code{"met_mlp_architecture"}.
}
\description{
Creates a multilayer perceptron architecture object from a list of dense
layers.
}
\examples{
architecture <- mlp_architecture(
  layers = list(
    dense_layer(10, activation = "relu"),
    dense_layer(1, activation = "linear")
  )
)
architecture
}
