% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{new_optimizer}
\alias{new_optimizer}
\title{Create an Optimizer Object}
\usage{
new_optimizer(name, type, parameters = list())
}
\arguments{
\item{name}{A character string specifying the optimizer name.}

\item{type}{A character string specifying the optimizer type.}

\item{parameters}{A list of optimizer-specific parameters.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Internal helper for constructing optimizer objects.
}
\keyword{internal}
