% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_abc}
\alias{optimizer_abc}
\title{Artificial Bee Colony Optimizer}
\usage{
optimizer_abc(colony_size = 30, max_iter = 100, limit = NULL)
}
\arguments{
\item{colony_size}{Total colony size. Half of the colony is used as employed
bees and half as onlooker bees.}

\item{max_iter}{Maximum number of iterations.}

\item{limit}{Number of unsuccessful trials before a food source is abandoned.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates an Artificial Bee Colony optimizer object for continuous
optimization problems.
}
\examples{
optimizer_abc()
}
\references{
Karaboga, D., and Basturk, B. (2007). A Powerful and Efficient Algorithm
for Numerical Function Optimization: Artificial Bee Colony (ABC) Algorithm.
Journal of Global Optimization, 39, 459--471.
doi:10.1007/s10898-007-9149-x

Karaboga, D., and Ozturk, C. (2009). Neural Networks Training by
Artificial Bee Colony Algorithm on Pattern Classification.
Neural Network World, 19(3), 279--292.
}
