% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_adam}
\alias{optimizer_adam}
\title{Adam Optimizer}
\usage{
optimizer_adam(
  learning_rate = 0.001,
  beta1 = 0.9,
  beta2 = 0.999,
  epsilon = 1e-08,
  epochs = 100,
  batch_size = NULL
)
}
\arguments{
\item{learning_rate}{Learning rate.}

\item{beta1}{Exponential decay rate for the first moment estimates.}

\item{beta2}{Exponential decay rate for the second moment estimates.}

\item{epsilon}{Small positive constant for numerical stability.}

\item{epochs}{Number of training epochs.}

\item{batch_size}{Mini-batch size. If \code{NULL}, full-batch training is used.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates an Adam optimizer object.
}
\examples{
optimizer_adam()
}
\references{
Kingma, D. P., and Ba, J. (2015). Adam: A Method for Stochastic
Optimization. International Conference on Learning Representations.
}
