% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_gwo}
\alias{optimizer_gwo}
\title{Grey Wolf Optimizer}
\usage{
optimizer_gwo(pop_size = 30, max_iter = 100, a_start = 2, a_end = 0)
}
\arguments{
\item{pop_size}{Population size.}

\item{max_iter}{Maximum number of iterations.}

\item{a_start}{Initial value of the control parameter \code{a}.}

\item{a_end}{Final value of the control parameter \code{a}.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates a Grey Wolf Optimizer object for continuous optimization problems.
}
\examples{
optimizer_gwo()
}
\references{
Mirjalili, S., Mirjalili, S. M., and Lewis, A. (2014). Grey Wolf Optimizer.
Advances in Engineering Software, 69, 46--61.
doi:10.1016/j.advengsoft.2013.12.007

Mirjalili, S. (2015). How Effective is the Grey Wolf Optimizer in Training
Multi-Layer Perceptrons. Applied Intelligence, 43, 150--161.
doi:10.1007/s10489-014-0645-7
}
