% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_hybrid}
\alias{optimizer_hybrid}
\title{Hybrid Optimizer}
\usage{
optimizer_hybrid(
  global = optimizer_pso(),
  local = optimizer_adam(),
  strategy = "sequential"
)
}
\arguments{
\item{global}{A metaheuristic optimizer object.}

\item{local}{A gradient-based optimizer object.}

\item{strategy}{Hybrid training strategy. Currently \code{"sequential"} is used
as the default strategy.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates a hybrid optimizer object by combining a global optimizer and a
local optimizer.
}
\examples{
optimizer_hybrid(
  global = optimizer_pso(max_iter = 10),
  local = optimizer_adam(epochs = 10)
)
}
