% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_info.R
\name{optimizer_info}
\alias{optimizer_info}
\title{Get Optimizer Information}
\usage{
optimizer_info(optimizer)
}
\arguments{
\item{optimizer}{Character name of an optimizer or an optimizer object created
by functions such as \code{optimizer_pso()}, \code{optimizer_sboa()},
\code{optimizer_sgd()}, or \code{optimizer_adam()}.}
}
\value{
An object of class \code{"met_optimizer_info"}.
}
\description{
Returns basic information about an optimizer available in the metANN package.
}
\examples{
optimizer_info("pso")
optimizer_info("sboa")
optimizer_info(optimizer_adam())
}
