% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_sboa}
\alias{optimizer_sboa}
\title{Secretary Bird Optimization Algorithm Optimizer}
\usage{
optimizer_sboa(pop_size = 30, max_iter = 100)
}
\arguments{
\item{pop_size}{Population size.}

\item{max_iter}{Maximum number of iterations.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates a Secretary Bird Optimization Algorithm optimizer object for
continuous optimization problems.
}
\examples{
optimizer_sboa()
}
\references{
Fu, Y., Liu, D., Chen, J., and He, L. (2024). Secretary Bird Optimization
Algorithm: A New Metaheuristic for Solving Global Optimization Problems.
Artificial Intelligence Review, 57, 123.
doi:10.1007/s10462-024-10729-y

Dilber, B., and Ozdemir, A. F. (2026). A novel approach to training
feed-forward multi-layer perceptrons with recently proposed secretary bird
optimization algorithm. Neural Computing and Applications, 38(5).
doi:10.1007/s00521-026-11874-x
}
