% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_sgd}
\alias{optimizer_sgd}
\title{Stochastic Gradient Descent Optimizer}
\usage{
optimizer_sgd(learning_rate = 0.01, epochs = 100, batch_size = NULL)
}
\arguments{
\item{learning_rate}{Learning rate.}

\item{epochs}{Number of training epochs.}

\item{batch_size}{Mini-batch size. If \code{NULL}, full-batch training is used.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates a stochastic gradient descent optimizer object.
}
\examples{
optimizer_sgd()
}
\references{
Robbins, H., and Monro, S. (1951). A Stochastic Approximation Method.
The Annals of Mathematical Statistics, 22(3), 400--407.
doi:10.1214/aoms/1177729586
}
