% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_tlbo}
\alias{optimizer_tlbo}
\title{Teaching-Learning-Based Optimization Optimizer}
\usage{
optimizer_tlbo(pop_size = 30, max_iter = 100)
}
\arguments{
\item{pop_size}{Population size.}

\item{max_iter}{Maximum number of iterations.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates a Teaching-Learning-Based Optimization optimizer object for
continuous optimization problems.
}
\examples{
optimizer_tlbo()
}
\references{
Rao, R. V., Savsani, V. J., and Vakharia, D. P. (2011).
Teaching-Learning-Based Optimization: A Novel Method for Constrained
Mechanical Design Optimization Problems. Computer-Aided Design, 43,
303--315. doi:10.1016/j.cad.2010.12.015
}
