% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_woa}
\alias{optimizer_woa}
\title{Whale Optimization Algorithm Optimizer}
\usage{
optimizer_woa(pop_size = 30, max_iter = 100, a_start = 2, a_end = 0, b = 1)
}
\arguments{
\item{pop_size}{Population size.}

\item{max_iter}{Maximum number of iterations.}

\item{a_start}{Initial value of the control parameter \code{a}.}

\item{a_end}{Final value of the control parameter \code{a}.}

\item{b}{Constant defining the spiral shape in the bubble-net mechanism.}
}
\value{
An object of class \code{"met_optimizer"}.
}
\description{
Creates a Whale Optimization Algorithm optimizer object for continuous
optimization problems.
}
\examples{
optimizer_woa()
}
\references{
Mirjalili, S., and Lewis, A. (2016). The Whale Optimization Algorithm.
Advances in Engineering Software, 95, 51--67.
doi:10.1016/j.advengsoft.2016.01.008
}
