% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_optimize.R
\name{plot.met_optimize_result}
\alias{plot.met_optimize_result}
\title{Plot Optimization Convergence}
\usage{
\method{plot}{met_optimize_result}(x, log = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"met_optimize_result"}.}

\item{log}{Logical. If \code{TRUE}, the y-axis is shown on a logarithmic scale.
Only positive convergence values can be displayed on a log scale.}

\item{...}{Additional graphical arguments passed to \code{plot()}.}
}
\value{
The input object invisibly.
}
\description{
Plots the convergence curve of a metANN optimization result.
}
