% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metann.R
\name{plot.metann}
\alias{plot.metann}
\title{Plot a metANN Model}
\usage{
\method{plot}{metann}(x, ...)
}
\arguments{
\item{x}{A fitted metANN model.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
The input object invisibly.
}
\description{
Plot a metANN Model
}
