% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metann.R
\name{predict.metann}
\alias{predict.metann}
\title{Predict with a metANN Model}
\usage{
\method{predict}{metann}(
  object,
  newdata,
  type = c("response", "prob", "class"),
  threshold = 0.5,
  ...
)
}
\arguments{
\item{object}{A fitted object of class \code{"metann"}.}

\item{newdata}{New data used for prediction. For formula-based models, this
should be a data frame. For x-y models, this should be a numeric matrix or
numeric data frame.}

\item{type}{Prediction type. For regression models, \code{"response"} returns
numeric predictions. For classification models, \code{"class"} returns predicted
class labels, \code{"prob"} returns predicted probabilities, and \code{"response"}
returns the default response, which is class labels.}

\item{threshold}{Classification threshold for binary classification.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector, matrix, or factor depending on the task and
prediction type.
}
\description{
Generates predictions from a fitted metANN model.
}
