% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{print.met_dense_layer}
\alias{print.met_dense_layer}
\title{Print a Dense Layer}
\usage{
\method{print}{met_dense_layer}(x, ...)
}
\arguments{
\item{x}{A dense layer object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object invisibly.
}
\description{
Print a Dense Layer
}
