% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{print.met_mlp_architecture}
\alias{print.met_mlp_architecture}
\title{Print an MLP Architecture}
\usage{
\method{print}{met_mlp_architecture}(x, ...)
}
\arguments{
\item{x}{An MLP architecture object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object invisibly.
}
\description{
Print an MLP Architecture
}
