% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_optimize.R
\name{print.met_optimize_result}
\alias{print.met_optimize_result}
\title{Print a metANN Optimization Result}
\usage{
\method{print}{met_optimize_result}(x, ...)
}
\arguments{
\item{x}{A metANN optimization result object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object invisibly.
}
\description{
Print a metANN Optimization Result
}
