% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{print.met_optimizer}
\alias{print.met_optimizer}
\title{Print a metANN Optimizer}
\usage{
\method{print}{met_optimizer}(x, ...)
}
\arguments{
\item{x}{A metANN optimizer object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object invisibly.
}
\description{
Print a metANN Optimizer
}
