% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_info.R
\name{print.met_optimizer_info}
\alias{print.met_optimizer_info}
\title{Print Optimizer Information}
\usage{
\method{print}{met_optimizer_info}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"met_optimizer_info"}.}

\item{...}{Additional arguments.}
}
\value{
The input object invisibly.
}
\description{
Print Optimizer Information
}
