% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metann.R
\name{print.metann}
\alias{print.metann}
\title{Print a metANN Model}
\usage{
\method{print}{metann}(x, ...)
}
\arguments{
\item{x}{A metANN model object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object invisibly.
}
\description{
Print a metANN Model
}
