% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print.metann_evaluation}
\alias{print.metann_evaluation}
\title{Print metANN Evaluation Results}
\usage{
\method{print}{metann_evaluation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"metann_evaluation"}.}

\item{...}{Additional arguments.}
}
\value{
The input object invisibly.
}
\description{
Print metANN Evaluation Results
}
