% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_optimize.R
\name{summary.met_optimize_result}
\alias{summary.met_optimize_result}
\title{Summarize a metANN Optimization Result}
\usage{
\method{summary}{met_optimize_result}(object, ...)
}
\arguments{
\item{object}{A metANN optimization result object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A list containing the main optimization results.
}
\description{
Summarize a metANN Optimization Result
}
