% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metann.R
\name{summary.metann}
\alias{summary.metann}
\title{Summarize a metANN Model}
\usage{
\method{summary}{metann}(object, ...)
}
\arguments{
\item{object}{A metANN model object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A list containing model summary information.
}
\description{
Summarize a metANN Model
}
