% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_optimize.R
\name{tlbo_optimize}
\alias{tlbo_optimize}
\title{Teaching-Learning-Based Optimization Engine}
\usage{
tlbo_optimize(fn, lower, upper, optimizer, verbose = TRUE, ...)
}
\arguments{
\item{fn}{Objective function.}

\item{lower}{Numeric vector of lower bounds.}

\item{upper}{Numeric vector of upper bounds.}

\item{optimizer}{A TLBO optimizer object.}

\item{verbose}{Logical. If \code{TRUE}, progress information is printed.}

\item{...}{Additional arguments passed to \code{fn}.}
}
\value{
An object of class \code{"met_optimize_result"}.
}
\description{
Internal implementation of Teaching-Learning-Based Optimization for
continuous optimization.
}
\keyword{internal}
